\name{IC.Baily1}
\alias{IC.Baily1}
\title{IC.Baily1}
\usage{
IC.Baily1()
}
\description{
IC.Baily1, world!'.
}
\arguments{
  \item{datos}{a vector of the counts.}
  \item{level}{the confidence level of the interval.}
}
\references{Correa-Morales, J. C. (2022) Analysis of CATegoriacal Data Via GSK and Simulation Using R: Lecture Notes. Escuela de Estadistica. Facultad de Ciencias. Universidad Nacional de Colombia. Sede Medellin.}
\examples{
IC.Baily1()
}
