\name{IC.F}
\alias{IC.F}
\title{IC.F}
\usage{
IC.F()
}
\description{
Prints 'Hello, world!'.
}
\arguments{
  \item{datos}{a vector of the counts.}
  \item{level}{the confidence level of the interval.}
}
\references{Correa-Morales, J. C. (2022) Analysis of CATegoriacal Data Via GSK and Simulation Using R: Lecture Notes. Escuela de Estadistica. Facultad de Ciencias. Universidad Nacional de Colombia. Sede Medellin.}
\examples{
IC.F()
}
