\name{TCL.II}
\alias{TCL.II}
\title{Multinomial confidence interval base on TCL}
\usage{
TCL.II()
}
\description{
Computes a simultaneous confidence region for the parameters of a multinomial parameter base on the Central Limit Theorem.
}
\arguments{
  \item{datos}{a vector of the counts.}
  \item{level}{the confidence level of the interval.}
}
\examples{
TLC.II(c(20,40,10))
}
\references{Correa-Morales, J. C. (2022) Analysis of CATegoriacal Data Via GSK and Simulation Using R: Lecture Notes. Escuela de Estadistica. Facultad de Ciencias. Universidad Nacional de Colombia. Sede Medellin.}

