\name{p.pois.trunc}
\alias{p.pois.trunc}
\title{p.pois.trunc}
\usage{
p.pois.trunc()
}
\description{
Prints 'Hello, world!'.
}
\arguments{
  \item{datos}{a vector of the counts.}
  \item{level}{the confidence level of the interval.}
}
\examples{
p.pois.trunc()
}
\references{Correa-Morales, J. C. (2022) Analysis of CATegoriacal Data Via GSK and Simulation Using R: Lecture Notes. Escuela de Estadistica. Facultad de Ciencias. Universidad Nacional de Colombia. Sede Medellin.}

